2️⃣🧮 update queries & stream to use double instead of int

+++ packages/backend/dart/lib/tables/_real_time_enabled/existing_collaborations/queries.dart
@@ -110,7 +110,7 @@ class ExistingCollaborationsQueries extends CollaborativeQueries {
   }
 
   Future<List> updateCurrentPhases(
-    int newPhase, {
+    double newPhase, {
     bool shouldEditCollaboratorsInfo = false,
   }) async {
     await ensureActiveCollaboratorInfo();
